Fix web build broken by use of react-native-skia
