+++ apps/tlon-mobile/package.json
@@ -57,7 +57,7 @@
     "@react-navigation/native-stack": "^6.9.13",
     "@shopify/flash-list": "1.6.3",
     "@tamagui/react-native-media-driver": "^1.116.12",
-    "@shopify/react-native-skia": "^1.3.9",
+    "@shopify/react-native-skia": "^1.5.3",
     "@tanstack/react-query": "~5.32.1",
     "@tloncorp/app": "workspace:*",
     "@tloncorp/editor": "workspace:*",
