+++ packages/ui/package.json
@@ -54,7 +54,7 @@
   },
   "peerDependencies": {
     "@10play/tentap-editor": "~0.5.21",
-    "@shopify/react-native-skia": "^1.3.9",
+    "@shopify/react-native-skia": "^1.5.3",
     "@urbit/sigil-js": "^2.2.0",
     "expo-av": "~13.10.5",
     "expo-image": "*",
