+++ packages/ui/src/components/DrawingInput/index.ts
@@ -1,0 +1,1 @@
+export { StandaloneDrawingInput } from './DrawingInput';
