+++ packages/ui/src/components/DrawingInput/shared.ts
@@ -1,0 +1,7 @@
+import type { ComponentType } from 'react';
+
+import type { DraftInputContext } from '../draftInputs';
+
+export type DrawingInputComponent = ComponentType<{
+  draftInputContext: DraftInputContext;
+}>;
